refactor: split image parsers into platform-specific implementations for web and IO
