+++ packages/stac/lib/src/parsers/widgets/stac_decoration_image/stac_decoration_image.dart
@@ -1,10 +1,11 @@
-import 'dart:io';
-
 import 'package:flutter/material.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:stac/src/parsers/widgets/stac_double/stac_double.dart';
 import 'package:stac/src/parsers/widgets/stac_rect/stac_rect.dart';
 import 'package:stac/stac.dart';
+
+export 'stac_decoration_image_parser_web.dart'
+    if (dart.library.io) 'stac_decoration_image_parser_io.dart';
 
 part 'stac_decoration_image.freezed.dart';
 part 'stac_decoration_image.g.dart';
@@ -30,39 +31,4 @@ abstract class StacDecorationImage with _$StacDecorationImage {
 
   factory StacDecorationImage.fromJson(Map<String, dynamic> json) =>
       _$StacDecorationImageFromJson(json);
-}
-
-extension StacDecorationImageParser on StacDecorationImage? {
-  DecorationImage? get parse {
-    if (this?.src == null) return null;
-
-    late ImageProvider image;
-    switch (this?.imageType) {
-      case StacDecorationImageType.network:
-        image = NetworkImage(this?.src ?? '');
-        break;
-      case StacDecorationImageType.file:
-        image = FileImage(File(this?.src ?? ''));
-        break;
-      case StacDecorationImageType.asset:
-        image = AssetImage(this?.src ?? '');
-        break;
-      default:
-        return null;
-    }
-
-    return DecorationImage(
-      image: image,
-      fit: this?.fit,
-      alignment: this?.alignment.value ?? Alignment.center,
-      centerSlice: this?.centerSlice?.parse,
-      repeat: this?.repeat ?? ImageRepeat.noRepeat,
-      matchTextDirection: this?.matchTextDirection ?? false,
-      scale: this?.scale.parse ?? 1.0,
-      opacity: this?.opacity.parse ?? 1.0,
-      filterQuality: this?.filterQuality ?? FilterQuality.low,
-      invertColors: this?.invertColors ?? false,
-      isAntiAlias: this?.isAntiAlias ?? false,
-    );
-  }
 }
